+++ modules/apps/foundation/portal-security-sso/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/events/CookieLogoutProcessor.java
@@ -17,8 +17,6 @@ package com.liferay.portal.security.sso.token.internal.events;
 import com.liferay.portal.kernel.util.CookieKeys;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.security.sso.token.events.LogoutProcessor;
-import com.liferay.portal.security.sso.token.events.LogoutProcessorType;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
