+++ modules/apps/foundation/portal-security-sso/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/events/TokenLogoutAction.java
@@ -24,9 +24,7 @@ import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.security.sso.token.configuration.TokenConfiguration;
-import com.liferay.portal.security.sso.token.constants.TokenConstants;
-import com.liferay.portal.security.sso.token.events.LogoutProcessor;
-import com.liferay.portal.security.sso.token.events.LogoutProcessorType;
+import com.liferay.portal.security.sso.token.internal.constants.TokenConstants;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
