+++ modules/apps/foundation/portal-security-sso/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/module/configuration/definition/TokenCompanyServiceConfigurationPidMapping.java
@@ -16,7 +16,7 @@ package com.liferay.portal.security.sso.token.internal.module.configuration.defi
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
 import com.liferay.portal.security.sso.token.configuration.TokenConfiguration;
-import com.liferay.portal.security.sso.token.constants.TokenConstants;
+import com.liferay.portal.security.sso.token.internal.constants.TokenConstants;
 
 import org.osgi.service.component.annotations.Component;
