+++ modules/apps/foundation/portal-security-sso/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/security/auth/SessionTokenRetriever.java
@@ -14,9 +14,6 @@
 
 package com.liferay.portal.security.sso.token.internal.security.auth;
 
-import com.liferay.portal.security.sso.token.security.auth.TokenLocation;
-import com.liferay.portal.security.sso.token.security.auth.TokenRetriever;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
