+++ modules/apps/foundation/portal-security-sso/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/verify/ShibbolethCompanySettingsVerifyProcess.java
@@ -20,9 +20,9 @@ import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.settings.SettingsFactory;
 import com.liferay.portal.kernel.util.PrefsProps;
 import com.liferay.portal.kernel.util.SetUtil;
-import com.liferay.portal.security.sso.token.constants.LegacyTokenPropsKeys;
-import com.liferay.portal.security.sso.token.constants.TokenConfigurationKeys;
-import com.liferay.portal.security.sso.token.constants.TokenConstants;
+import com.liferay.portal.security.sso.token.internal.constants.LegacyTokenPropsKeys;
+import com.liferay.portal.security.sso.token.internal.constants.TokenConfigurationKeys;
+import com.liferay.portal.security.sso.token.internal.constants.TokenConstants;
 import com.liferay.portal.verify.BaseCompanySettingsVerifyProcess;
 import com.liferay.portal.verify.VerifyProcess;
